Experimental

+++ Class/pipe.py
@@ -25,10 +25,10 @@ class Pipe:
         if not ip: return False
         return True
 
-    def prepare(self,server,Filter=True,delete=False):
+    def prepare(self,server,threading=False,Filter=True,delete=False,ignorelist=[]):
         print("---",server,"Preparing","---")
         #Check if v6 only
-        serverSuffix = ""
+        serverSuffix,threads = "",[]
         if self.checkResolve(server) is False and self.checkResolve(server+"v6") is True:
             print("Switching",server,"to v6 only")
             serverSuffix ="v6"
@@ -42,8 +42,11 @@ class Pipe:
             if client.endswith("Serv") and Filter == True or Filter == False:
                 #Stop Server
                 print("Stopping",client.replace("Serv",""),"on",server)
-                self.cmd(server+serverSuffix,'systemctl stop wg-quick@'+client+' && systemctl disable wg-quick@'+client)
-                if delete == True:
+                if threading and client not in ignorelist:
+                    threads.append(Thread(target=self.cmd, args=([server+serverSuffix,'systemctl stop wg-quick@'+client+' && systemctl disable wg-quick@'+client])))
+                elif client not in ignorelist:
+                    self.cmd(server+serverSuffix,'systemctl stop wg-quick@'+client+' && systemctl disable wg-quick@'+client)
+                if delete == True and client not in ignorelist:
                     self.cmd(server+serverSuffix,'rm -f /etc/wireguard/'+client+".conf")
                 #Stop Client
                 v6 = 'v6' if client.endswith("v6Serv") else ''
@@ -54,26 +57,44 @@ class Pipe:
                 else:
                     suffix = ""
                 print("Stopping",self.targets['prefix']+server+v6,"on",client+suffix)
-                self.cmd(client+suffix,'systemctl stop wg-quick@'+self.targets['prefix']+server+v6+' && systemctl disable wg-quick@'+self.targets['prefix']+server+v6)
-                if delete == True:
+                if threading and client not in ignorelist:
+                    threads.append(Thread(target=self.cmd, args=([client+suffix,'systemctl stop wg-quick@'+self.targets['prefix']+server+v6+' && systemctl disable wg-quick@'+self.targets['prefix']+server+v6])))
+                elif client not in ignorelist:
+                    self.cmd(client+suffix,'systemctl stop wg-quick@'+self.targets['prefix']+server+v6+' && systemctl disable wg-quick@'+self.targets['prefix']+server+v6)
+                if delete == True and client not in ignorelist:
                     self.cmd(client+suffix,'rm -f /etc/wireguard/'+self.targets['prefix']+server+v6+".conf")
+        if threading: self.lunchThreads(threads)
 
     def clean(self):
-        threads = []
+        threads,ignoreList = [],[]
         answer = input("Use Threading? (y/n): ")
+        ignore = input("Any nodes to ignore? (Name,Name../n): ")
+        if ignore != "n":
+            ignoreList = ignore.split(",")
         for server,data in self.targets['servers'].items():
+            if server in ignoreList: continue
             if answer != "y":
-                self.prepare(server,False,True)
+                self.prepare(server,False,False,True,ignoreList)
             else:
-                threads.append(Thread(target=self.prepare, args=([server,False,True])))
+                threads.append(Thread(target=self.prepare, args=([server,False,False,True,ignoreList])))
         if answer == "y": self.lunchThreads(threads)
+
+    def check(self):
+        for server,data in self.targets['servers'].items():
+            print("---",server,"Checking","---")
+            suffix = ""
+            if self.checkResolve(server) is False and self.checkResolve(server+"v6") is True:
+                print("Switching",server,"to v6 only")
+                suffix = "v6"
+            nics = self.cmd(server+suffix,'ip addr show')[0]
+            print(nics)
 
     def shutdown(self):
         threads = []
         answer = input("Use Threading? (y/n): ")
         for server,data in self.targets['servers'].items():
             if answer != "y":
-                self.prepare(server,False)
+                self.prepare(server,False,False)
             else:
                 threads.append(Thread(target=self.prepare, args=([server,False])))
         if answer == "y": self.lunchThreads(threads)
@@ -81,7 +102,7 @@ class Pipe:
     def lunchThreads(self,threads):
         for thread in threads:
             thread.start()
-            time.sleep(0.1)
+            time.sleep(0.2)
         for thread in threads:
             thread.join()
 
@@ -127,9 +148,10 @@ class Pipe:
         print('Done',client,'on',server)
 
     def run(self):
-        start = 4
+        threading,start = False,4
         crossConnect,clients,threads = [],[],[]
         answer = input("Use Threading? (y/n): ")
+        if answer == "y": threading = True
         print("Launching")
         time.sleep(3)
         for server,data in self.targets['servers'].items():
@@ -138,7 +160,7 @@ class Pipe:
                 self.targets['servers'][server]['basePort'] = port = random.randint(1500, 55000)
             else:
                 port = data['basePort']
-            self.prepare(server)
+            self.prepare(server,threading)
             print("---",server,"Deploying","---")
             #Check if v6 only
             v6only,suffix = False,""
